Move `Rescheduler` into a separate package-access class
